youtube-pipeline/requirements.txt
# YouTube 自動化パイプライン 依存パッケージ

# Claude API
anthropic>=0.40.0

# YouTube Data API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# RSS / Web取得
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0

# 音声生成
openai>=1.12.0          # OpenAI TTS fallback用
gTTS>=2.4.0             # Google TTS fallback

# 映像・画像処理
Pillow>=10.1.0          # サムネイル生成
# ffmpeg-python>=0.2.0  # pip経由 (別途 ffmpeg バイナリ必要)

# データ処理
pyyaml>=6.0.1
python-dotenv>=1.0.0
jinja2>=3.1.2           # テンプレート

# ユーティリティ
tqdm>=4.66.0            # プログレスバー
tenacity>=8.2.3         # リトライ処理
